feat: 반복 소비 내역 삭제 Related to: #2
