+++ src/main/java/com/umc5th/muffler/domain/routine/controller/RoutineController.java
@@ -1,13 +1,11 @@
 package com.umc5th.muffler.domain.routine.controller;
 
 import com.umc5th.muffler.domain.routine.dto.*;
+import com.umc5th.muffler.domain.routine.service.RoutineCreateService;
 import com.umc5th.muffler.domain.routine.service.RoutineService;
 import com.umc5th.muffler.global.response.Response;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
@@ -16,20 +14,28 @@ import javax.validation.Valid;
 @RequestMapping("/routine")
 public class RoutineController {
 
+    private final RoutineCreateService routineCreateService;
     private final RoutineService routineService;
 
     @PostMapping("/weekly")
     public Response<Void> addWeeklyRoutine(@RequestBody @Valid AddWeeklyRoutineRequest request) {
 
-        routineService.addWeeklyRoutine(request);
-        routineService.addPastExpenses(request);
+        routineCreateService.addWeeklyRoutine(request);
+        routineCreateService.addPastExpenses(request);
         return Response.success();
     }
 
     @PostMapping("/monthly")
     public Response<Void> addMonthlyRoutine(@RequestBody @Valid AddMonthlyRoutineRequest request) {
 
-        routineService.addMonthlyRoutine(request);
+        routineCreateService.addMonthlyRoutine(request);
+        return Response.success();
+    }
+
+    @DeleteMapping("/{routineId}")
+    public Response<Void> deleteRoutine(@PathVariable Long routineId, @RequestParam Long memberId) {
+
+        routineService.delete(routineId, memberId);
         return Response.success();
     }
 }
